@@ -1,7 +1,7 @@
 {
   "name": "solitude",
   "private": true,
-  "version": "0.1.1",
+  "version": "0.2.0",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
@@ -9,16 +9,17 @@
     "tauri": "tauri"
   },
   "dependencies": {
-    "@tauri-apps/api": "^1.0.0-rc.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.2"
+    "@tauri-apps/api": "^1.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-use": "^17.4.0"
   },
   "devDependencies": {
-    "@tauri-apps/cli": "^1.0.0-rc.7",
-    "@vitejs/plugin-react": "^1.0.7",
-    "vite": "^2.8.0",
-    "vite-plugin-windicss": "^1.8.3",
-    "windicss": "^3.5.1"
+    "@tauri-apps/cli": "^1.0.5",
+    "@vitejs/plugin-react": "^2.0.0",
+    "vite": "^3.0.4",
+    "vite-plugin-windicss": "^1.8.7",
+    "windicss": "^3.5.6"
   }
 }
