@@ -1,8 +1,11 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-import WindiCSS from 'vite-plugin-windicss'
+import { defineConfig } from "vite"
+import react from "@vitejs/plugin-react"
+import WindiCSS from "vite-plugin-windicss"
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(),WindiCSS()]
+  plugins: [react(), WindiCSS()],
+  server: {
+    port: 3000,
+  },
 })
